[CMAKE] Move .h files transformed from .ui to the form subdirectory
Summary:
This diff is an alternative to D2750.

This is a prerequisite for PR11651, as all the qt headers will be
included with their path relative to `src/`.

Example: `#include "ui_intro.h` will become `#include
<qt/forms/ui_intro.h>`.

The `ui_*.h` files generated from the `*.ui` files are generated by
`qt5_wrap_ui` in the `CMakeLists.txt` file directory, wherever the
source `*.ui` files are located.

This diff updates then restores the `CMAKE_CURRENT_BINARY_DIR` in order
to make CMake behave as if the `CMakeLists.txt` file was located into
the `qt/forms` directory.

Test Plan:
  mkdir buildcmake && cd buildcmake
  cmake -GNinja ..
  ninja check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D2786