Remove Windows 32 bit build
Summary:
The win32 target will no longer be part of our release nor officially
supported.

It is still possible for users to run the build by themselves, but the
instructions are removed from the doc in order to reflect the
"unsupported" status of this target.

Backport of core [[https://github.com/bitcoin/bitcoin/pull/15939 | PR15939]].

Depends on D5610.

Test Plan:
Run the Windows Gitian build. Check the win32 binaries are no longer
part of the output.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5612

Conflicts:
	doc/build-windows.md
	doc/release-notes.md