[CMAKE] Disable OpenGL in Qt static build for OSX
Summary:
A bug in the CMake files distributed by Qt prevent building with static
OpenGL on OSX (hard linked dependency to the framework in the Qt5Gui
dependencies). To workaround the issue, the static Qt is built without
the OpenGL support.

This fixes the `bitcoin-qt` cross build for OSX.

Depends on D3017 and D3033 (for the test plan).

Test Plan:
  mkdir buildcmake && cd buildcmake
  cmake -GNinja -DCMAKE_TOOLCHAIN_FILE=../cmake/platforms/OSX.cmake ..
  ninja
Run the generated src/qt/bitcoin-qt on OSX >= 10.8 (tested on OSX Sierra
10.12.6).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3034