Have gArgs handle printing help
Summary: This is a remake of Core PR13190 . Because the parameter set we have differs from core, I ended up having to redo most of it, even though the ogic is the same.

Test Plan:
  make check
  arc lint

Make sure that the doc linter is still happy.

  BUILDDIR=~/bitcoin-abc/build ../contrib/devtools/gen-manpages.sh

Make sure, that even though some arguments have been shuffled around, the doc is still good.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D3058