remove SCRIPT_ENABLE_SCHNORR flag and clean up tests
Summary:
This is a backing-out of D2430, which added an impotent
SCRIPT_ENABLE_SCHNORR flag along with a bunch of duplicate tests. Now that
the flag is once again impotent, we remove it.

In addition to removing those tests, there were numerous other script_tests
cases that are now redundant as well, also removed in this diff.

Depends on D3331

Test Plan: `make check`

Reviewers: deadalnix, Mengerian, #bitcoin_abc

Reviewed By: deadalnix, Mengerian, #bitcoin_abc

Subscribers: jasonbcox

Maniphest Tasks: T653

Differential Revision: https://reviews.bitcoinabc.org/D3332