[LINTER] Add a spell checker to arcanist
Summary:
This is not the same spell checker that the one used by core, but it
avoids adding another dependency and it runs faster. The dictionary is
also smaller but easier to maintain.

The dictionary is the one shipped with arcanist in which I added with the errors found in the codebase by core's linter that were not catched.

Test Plan:
Remove the line:
  "(^test/lint/dictionary/)"
from the `exclude` array of the `spelling` linter in the `.arclint`
file, as well as the trailing comma on the preceding line.
  arc lint -- test/lint/dictionary/english.json
Check the linter is linting its own dictionary.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3217