Replumb ibd.sh to prepare for better post-IBD checks
Summary:
With the current design of ibd.sh, exiting the IBD subshell early (in the
future case of adding additional tests/checks) would result in bitcoind not exiting
as expected.  This patch replumbs ibd.sh such that failures in the IBD subshell or
bitcoind properly fail the script as expected.

Post IBD checks will be added in an upcoming diff.

Test Plan: TeamCity IBD

Reviewers: #bitcoin_abc, deadalnix, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D3359