Use a unique index for the running jobs in case of duplicated names
Summary:
The core's PR10374 allows to run the same test several times with a
single call to `test_runner.py`, by duplicating the test name in the
arguments list. The test_runner.py code has diverged from core, and some
changes will prevent from running duplicated tests.

The `test_runner.py` script maintains a set of running jobs, indexed by
their test name. This diffs changes the set to a list, and uses a tuple
(test number, test name) as the key. This also requires to pass the
number to the `TestResult`  class in order to rebuild the key when the
job needs to be removed from the list.

Test Plan:
  test_runner.py

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3367