Speedup coinselector_tests by using a dummy WalletDBWrapper when apropriate
Summary: Since the addition of AddToWallet to add_coin, some test became significantly slower due to an increase in operation done on the mock DB. Using a dummy DB when the mock is not required speeds things up dramatically.

Test Plan:
  perf -g -- src/test/test_bitcoin -t coinselector_tests
  perf report

Make sure that the majority of the time isn't spent in BDB anymore.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D3387