Move make check test logs out of tree
Summary:
Thanks to @deadalnix for pointing out the issue.

Follow-up for D3535.

When running `make check` all the tests output a `<test_file>.log` file
to the `src/test/` directory (in tree).
This diffs moves these logs to the `<build_dir>/src/test/` directory
(out of tree) with the name `<test_file>.test.log`.

Test Plan:
  mkdir build && cd build
  make check
Ensure the test logs are in the correct place.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3591