[LINTER] Add a linter to replace `unsigned char` with `uint8_t`
Summary:
This is a recurrent issue in reviews, this diffs provides an autofix
linter for this.
It is named afted stdint so that other conversion can be lated added if
needed.

Test Plan:
  arc lint --everything # No warning

Edit `src/init.cpp` to add the line:
  unsigned char foo;
anywhere in the file, then run:
  arc lint -- src/init.cpp
The linter should catch the line and suggest a fix.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3592