Use functional tests timings from the source directory
Summary:
The current behavior is to read timing from the `timing.json` file
location in the build directory is it exists.
But this file is updated every time the functional test suite is run,
which makes it impossible to predict which test will run as part of the
`--extended` call and which one will run as part as a standard call.
This diffs change this behavior so that the `timing.json` file from the
source directory is always used as the timings source to determine which
test to run.

Thanks @markblundeberg for reporting the issue.

Test Plan:
  ./test/functional/test_runner.py
Change the timing for a previously ran test to a duration > 40s in the
`test/functional/timing.json` file.
  ./test/functional/test_runner.py
The test should not run anymore.

Reviewers: #bitcoin_abc, deadalnix, markblundeberg

Reviewed By: #bitcoin_abc, markblundeberg

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D3595