[script_tests] add tests for pubkey encoding with strictenc + nullfail
Summary:
During discussion of D3473 an optimization for the NULLFAIL check was
proposed. If implemented naively with bail-early mechanics, it could
lead to an inadvertent consensus change due to interaction with
STRICTENC. From what I can tell this was not covered by test vectors.

Depends on D3612.

Test Plan: `make check`

Reviewers: deadalnix, jasonbcox, Fabien, Mengerian, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D3613