Bump automatic replay protection to May 2020 upgrade
Summary:
This returns main chain to ForkID of 0 after the Nov 2019
upgrade, and implements automatic replay protection for
the May 2020 upgrade.

For reference see last HF:
https://reviews.bitcoinabc.org/D2376

Test Plan:
Replay protection is covered by unit tests, however
it is very important to manually double check that the timestamp
for activation is set correctly.

Reviewers: deadalnix, markblundeberg, jasonbcox, Fabien, #bitcoin_abc

Reviewed By: markblundeberg, Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D3868