Increase timeout on avalanche test.
Summary: This shouldn't be an issue, because the loop break as soon as the event is reached.

Test Plan:
Run the test and check there was no failure, and that it did not take forever to run.

As the issue cannot be reproduced reliably, this cannot ensure that this fixes it, but it should be enough.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D3962