[CI] Refactor the build configuration by moving out tests from the build
Summary:
This makes it easier to individually add and configure each step for
each configuration, and makes the whole build definition easier to read.

As a (good) side effect the functional tests are no longer run with the next
upgrade enabled with the `asan` and `ubsan` configurations, and use the
standard cutoff.

Depends on D3992 and D4014.

Test Plan: Run the CI builds/IBD and check the builds are green (ubsan excepted).

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3993