[CI] Workaround ubsan failure in functional tests
Summary:
This will allow to run the undefined behavior sanitizer on the
functional tests.
For now the `abc-p2p-compactblocks` functional test is failing with ubsan (a message is missed, with no log) and the node fails to shut down, which causes the build to hang until timeout. This test is excluded from the list in order to allow for the other tests to run with ubsan, waiting for a fix.

Depends on D3993.

Test Plan: Run the CI `build-ubsan` build configuration.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3994