[SECP256K1] Build java class files out of tree
Summary:
The Java files are compiled to the source tree, making it difficult to
restart a build. Fix this by passing the output directory to `javac`.

Test Plan:
  cd secp256k1
  ./autogen.sh
  mkdir build && cd build
  ../configure --enable-jni --enable-experimental --enable-module-ecdh
  make check-java
Check the `.class` files are located under the build directory
(`build/src/java/org/bitcoin/*.class`).
  make clean
Check the `.class` files are deleted.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4053