Add OpenSSL termios fix for musl libc
Summary:
```
Currently the version of OpenSSL included in the depends system won't
build on musl based systems because termio.h does not exist. The proper
header named termios.h does exist.

This PR adds a patch for OpenSSL to replace the termio.h header with
termios.h, which is the proper POSIX header as I understand it.

This is a known issue as TERMIOS (not TERMIO) should be the default, and
is fixed in later versions of OpenSSL. There is discussion on the
OpenSSL repo here: openssl/openssl#163

This has been fixed in OpenSSL.

This partly fixes #16925 and allows building Bitcoin on Alpine using the
depends system.
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/16926 | PR16926]].

Test Plan: Run the Linux Gitian builds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5651