[CMAKE] Install the man pages
Summary:
This adds support for the man pages installation.
The last directory is hardcoded to `man1`. This is correct as long as
the man files are hardcoded to section 1 only.

Test Plan:
For all platforms:
  mkdir -p buildcmake/install && cd install
  cmake -GNinja .. -DCMAKE_INSTALL_PREFIX="install" # Other options
depending on the platform
  ninja
  ninja install

Check that the man pages are installed under `install/share/man/man1`.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4147