[CMAKE] Consistently find and use the python interpreter
Summary:
Some of the python scripts rely on the shebang, some of them search for
the python program, the behavior is inconsistent across the files.

This diff makes use of `find_package` to get the python interpreter
executable path. This allow for failing with a meaningful error message
if python is not found or does not match the minimum version.

Depends on D4143.

Test Plan:
  mkdir buildcmake && cd buildcmake
  cmake -GNinja ..
  ninja check
  ninja check-functional
  ninja bench-bitcoin

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4144