Add support for Glibc version 2.28
Summary:
The Glibc backward compatibility feature handles up to version 2.27
(used by Ubuntu Bionic). This diffs adds support for the version 2.28
(used by Debian Buster).

There is an additional compatibility fix involved due to the update of
`fcntl()` and the introduction of the new `fcntl64()` function.
These functions are linked against their previous version through a
compatibility symbol.

Test Plan:
Cross build for all the linux platforms on an OS that ships Glibc 2.28
(tested on Debian Buster). Use both `autotools` and `cmake` for the
build.
Run the `symbol-check.py` script against the binaries and check there is
no complaint on `fcntl` or `fcntl64` symbol being from a too recent
Glibc.
Check the binaries run on OS with a `libc.so.6` version < 2.28

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4167