Fix --wipeignore and add message indicating if wipe options are set and successful
Summary:
The `--wipeignore` option was impossible to use because the constructor
would write the input for that parameter to `fWipeBan` instead.

Adds `fprintf()` statements for the `--wipeignore` and `--wipeban`
options.  This makes testings changes to these options easier and gives
the user some confimation that the options were used.

Test Plan:
  make
  ./bitcoin-seeder --wipeban --wipeignore
Output should be as below:
  Supporting whitelisted filters: 0x1,0x5,0x11,0x15
  No nameserver set. Not starting DNS server.
  Loading dnsseed.dat...Ban list wiped...Ignore list wiped...done
  Starting seeder...done

Reviewers: deadalnix, Fabien, jasonbcox, O1 Bitcoin ABC, #bitcoin_abc

Reviewed By: jasonbcox, O1 Bitcoin ABC, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D4249