[CMAKE] Build the OSX application bundle for bitcoin qt
Summary:
When building/cross building for OSX, create an application bundle that
can run natively on OSX.
This is a first step toward generating an installer for OSX.

Test Plan:
On OSX (or cross build):
  ninja BitcoinABC-Qt
Check there is an app bundle (`BitcoinABC-Qt.app`) created under
`src/qt`.
Check on OSX that the bundle has the correct icon, details and can be
run.

{F4082441}

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4290