[CMAKE] Migrate linux gitian build to cmake and ninja
Summary:
Replace `autotools` by `cmake` for the linux gitian builds, and use
`ninja` as a generator.

Note 1:
The `test_bitcoin` binary is no longer part of the gitian output.

Note 2: There are some more dependencies required.
This is because CMake is run for the native platform in order to build a
couple helper binaries for secp256k1 and univalue. CMake is run from the
project root and check for all the dependencies despite only a compiler
is required. This is a minor inconvenience which can be reworked later.

Test Plan:
Run the Linux gitian build twice.
Check the hashes are identical between the builds.
Run the binaries on all the platforms.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4356