[CMAKE] Move FDELT_TYPE declaration to config
Summary:
This is only used when the Glibc backward compatibility feature is
enabled, but moving the definition to the `bitcoin-config.h` instead of
the command line has no impact on code since the `glibc_compat.cpp` file
(the only callsite) is built conditionally.

This fixes a weird bug in Gitian builds, where passing the
`-DFDELT_TYPE="long int"` definition on the command line causes GCC to
fail. While being apparently correctly surrounded by the double quotes,
the compiler sees the `int` as a source file and aborts. This only
occurs when the script is read from the YAML descriptor during a Gitian
build, and cannot be reproduced by building from a shell.

Test Plan:
  cmake -GNinja .. -DENABLE_GLIBC_BACK_COMPAT=ON
  ninja

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4347