[GITIAN] Migrate OSX Gitian builds to CMake and Ninja
Summary:
Replace autotools by CMake for the OSX Gitian builds, and use
Ninja as a generator.

Note 1:
The test_bitcoin binary is no longer part of the Gitian output.

Note 2: There are some more dependencies required.
This is because CMake is run for the native platform in order to build a
couple helper binaries for secp256k1 and univalue. CMake is run from the
project root and check for all the dependencies despite only a compiler
is required. This is a minor inconvenience which can be reworked later.

Test Plan:
Run the OSX Gitian build twice.
Check the hashes are identical between the builds.
Run the binaries and installers (DMG).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4407