[CMAKE] Add the background image to the dist directory
Summary:
This diff generates the background image to be added to the OSX `dist`
directory. This background image will later be used as the background
for the OSX installer (DMG image).

This background file is expected to be a multi-resolution TIFF image. In
order to allow to change the package name easily, it is constructed from
a SVG file template which is configured at build time.

This SVG file is then converted to multiple DPI raster images, which are
finally concatenated into a single final background TIFF file.

This image transformation process requires several steps, and will
differ between a cross compilation and a native compilation due to the
different tool being available. This diff only support cross
compilation, native platform support will be added later.

Depends on D4380.

Test Plan:
  ninja osx-deploydir
Check the `background.tiff` file is added to the `dist` directory (under
a dedicated `.background` directory).
Open the TIFF file and check it contains 2 images with different (x2
ratio) resolutions.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4381