[CI] Run the thread sanitizer as part of the CI
Summary:
ASAN and UBSAN are already part of the CI, but TSAN was missing.
There is an exception in the functional test suite for
`wallet_multiwallet`. This is the exact same issue than the ASAN failure
(heap used after free).
It is possible that some tests end up being flacky due to the load
overhead from TSAN (timeouts may occur), but they should be fixed as
they appear on a case to case basis.

Test Plan:
  ABC_BUILD_NAME=build-tsan ./contrib/teamcity/build-configurations.sh

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4412