[CMAKE] Use a stripped binary to build the dist and DMG for OSX
Summary:
This diff creates a copy of the `bitcoin-qt` application bundle on OSX,
and strip the executable inside.
This stripped version is then use to build the DMG image.

Depends on D4437.

Test Plan:
Build for OSX:
  ninja osx-dmg
Check there is a `src/qt/stripped` directory containing a copy of the
bundle with a stripped binary inside.
Check the `dist` and DMG binary is stripped.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4438